feat: add windows support
We don't need to raise the file descriptor limit on windows.
