+++ go.mod
@@ -1,0 +1,8 @@
+module github.com/jbenet/hang-fds
+
+go 1.12
+
+require (
+	github.com/multiformats/go-multiaddr v0.0.4
+	github.com/multiformats/go-multiaddr-net v0.0.1
+)
