+++ rlimit_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package main
+
+func fdRaise(nn int) error {
+	return nil
+}
